requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
apscheduler==3.10.4
groq==0.8.0
requests==2.31.0
python-dotenv==1.0.1